Safari-611.1.10/LayoutTests/fast/dom/Geolocation/multiple-requests-expected.txt

Tests that Geolocation correctly handles multiple concurrent requests.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS oneShotCallbackInvoked is false
PASS position.coords.latitude is mockLatitude
PASS position.coords.longitude is mockLongitude
PASS position.coords.accuracy is mockAccuracy
PASS watchCallbackInvoked is false
PASS position.coords.latitude is mockLatitude
PASS position.coords.longitude is mockLongitude
PASS position.coords.accuracy is mockAccuracy
PASS successfullyParsed is true

TEST COMPLETE
